17.1.Agree to Terms Database Field{Adding the "Agreed Terms" Persisted Date Field: bin/console make:entity}
If you compare our old registration form and our new one, we're missing one annoying, piece: the "Agree to terms" checkbox, which, if you're like me, is just one of my favorite things in the world - right behind a fine wine or day at the beach.
Legally speaking, this field is important. So let's code it up correctly.

Adding the "Agreed Terms" Persisted Date Field
A few years ago, we might have added this as a simple unmapped checkbox field with some validation to make sure it was checked. But these days, to be compliant, we need to save the date the terms were agreed to.
Let's start by adding a new property for that! Find your terminal and run:

php bin/console make:entity
Update the User class and add a new field called agreedTermsAt. This will be a datetime field and it cannot be nullable in the database: we need this to always be set. Hit enter to finish.