17.4.Agree to Terms Database Field{Migrating Existing User Data}
To finish the change, make one more migration:

  php bin/console make:migration
Go check it out! Perfect! This gives us the last piece we need: changing the column back to NOT NULL, which will work because each existing user now has a real value for this field. Oh, but, for legal purposes, on a real site - it may not be proper to automatically set the agreed_terms_at for existing users. Yep, you've gotta check with a lawyer on that kind of stuff.
But from a database migration standpoint, this should fix everything! Run the last migration:

php bin/console doctrine:migrations:migrate
Excellent! Next: we have a CheckboxType field on the form... which is good at setting true/false values. And, we have an agreedTermsAt DateTime field on the User class. Somehow, those need to work together!