18.2.Agree to Terms Checkbox Field
Thanks to the mapped => false, the data from the checkbox does not affect our User object in any way when we submit. No problem: in SecurityController, let's handle it manually with if (true === $form['agreeTerms']->getData()). Wait... that looks redundant! We already have form validation that forces the box to be checked. You're totally right! I'm just being extra careful... ya know... for legal reasons.
Inside, we could call $user->setAgreedTermsAt() and pass the current date. Or, we can do something a bit cleaner. Find the setAgreedTermsAt() method and rename it to agreeTerms(), but with no arguments. Inside say $this->agreedTermsAt = new \DateTime().
This gives us a clean, meaningful method. In SecurityController, call that: $user->agreeTerms().
Ok team, let's try this. Refresh the page. Annoyingly, I still can't see the checkbox. Let's hack that for now: add a little extra padding on this div. There it is!
Register as geordi3@theenterprise.org, password engage, hit enter, and... yes! We know the datetime column was just set correctly in the database because it's required.
Here's the big takeaway: whenever you need a field on your form that doesn't exist on your entity, there may be a clever solution. But, if it's not obvious, make the field unmapped and add a little bit of glue code in your controller that does whatever you need.
Later, we'll discuss a third option: creating a custom model class for your form.