18.3.Agree to Terms Checkbox Field
Fixing your Fixtures
Before we move on, try to reload the fixtures:

  php bin/console doctrine:fixtures:load
It... explodes! Duh! I made the new agreedTermsAt field required in the database, but forgot to update it in the fixtures. No problem: open UserFixture. In the first block, add $user->agreeTerms(). Copy that, and do the same for the admin users.
Cool! Try it again:

  php bin/console doctrine:fixtures:load
And.... all better!
Next: let's fix the styling in our registration form by creating our very own form theme.