24.2.Data Transformer{The transform() Method}
The transform() Method

That's the whole point of transform()! This method is called. when the form is rendering: it takes the raw data for a field - in our case the User object that lives on the author property - and our job is to transform that into a representation that can be used for the form field. In other words, the email string.
First, if null is the value, just return an empty string. Next, let's add a sanity check: if (! $value instanceof User), then we, the developer, are trying to do something crazy. Throw a new LogicException() that says:

  The UserSelectTextType can only be used with User objects.
Finally, at the bottom, so nice, return $value - which we now know is a User object ->getEmail(). Let's rock! Move over, refresh and.... hello email address!