25.2.Custom Field: configureOptions() & Allowing Empty Input{Fixing Empty Value Case in the Data Transformer}
Fixing Empty Value Case in the Data Transformer
I want to talk more about field options because they can unlock some serious possibilities. But first, there is a teenie, tiny bug with our data transformer. Clear out the author text box and try to submit. Duh - disable HTML5 validation by adding the novalidate attribute. Hit update!
Oh! Our sanity validation still fails: User not found. That's not quite what we want. Instead of failing, our data transformer should probably just return null.
Go back to EmailToUserTransformer. In reverseTransform(), if  is empty, just return. So, if the field is submitted empty, null should be passed to setAuthor().
But, hmm... the problem now is that, while it's technically ok to call setAuthor() with a null argument, we want that field to be required!
Re-submit the form! Oof - an integrity constraint violation: it's trying to save to the database with null set as the author_id column. We purposely made this required in the database and this is a great example of... messing up! We forget to add an important piece of business validation: to make the author required.