25.3.Custom Field: configureOptions() & Allowing Empty Input{Fixing Empty Value Case in the Data Transformer}
No worries: open the Article class, find the $author field and, above it, add @Assert\NotNull(). Give it a message: Please set an author.
Try that again. Excellent! This is the behavior - and error - we expect.
Next: how could we make our custom field type behave differently if it was used in different forms? Like, what if in one form, we want the user to be able to enter any user's email address but in another form we only want to allow the user to enter the email address of an admin user. Let's learn more about the power of form field options.