29.1.Hooking up the AJAX Autocomplete{Adding a data-autocomplete-url Attribute}
We now have an endpoint that returns all users as JSON. And we have some autocomplete JavaScript that... ya know... autocompletes entries for us. I have a crazy idea: let's combine these two so that our autocomplete uses that Ajax endpoint!

Adding a data-autocomplete-url Attribute
First: inside of the JavaScript, we need to know what the URL is to this endpoint. We could hardcode this - I wouldn't judge you for doing that - this is a no-judgment zone. But, there is a simple, clean solution.
In AdminUtilityController, let's give our new route a name: admin_utility_users. Now, idea time: when we render the field, what if we added a "data" attribute onto the input field that pointed to this URL? If we did that, it would be super easy to read that from JavaScript.
Let's do it! In UserSelectTextType, add another attribute: how about data-autocomplete-url set to... hmm. We need to generate the URL to our new route. How do we generate a URL from inside of a service? Answer: by using the router service. Add a second argument to the constructor: RouterInterface $router. I'll hit Alt+Enter to add that property and set it.
Oh, and if you can't remember the type-hint to use, at least make sure that you remember that you can run:

php bin/console debug:autowiring
to see a full list of type-hints. By the way, in Symfony 4.2, this output will look a little bit different, but contains the same info. If you search for the word "route" without the e... cool! We have a few different type-hints, but they all return the same service anyways.
Now that we've injected the router, down below, use $this->router->generate() and pass it the new route name: admin_utility_users.
Let's check it out! Refresh, inspect that field and ... perfect! We have a shiny new data-autocomplete-url attribute.