32.2.Tweak your Form based on the Underlying Data{Accessing Data via $options}
Accessing Data via $options

How about: $article = $options['data'] ?? null;
If you don't know that syntax, it basically says that I want the $article variable to be equal to $options['data'] if it exists and is not null. But if it does not exist, set it to null. Let's dump that and make sure it's what we expect.
Refresh the new article page - yep - null. Try the edit page... there's the Article object. Now, we are dangerous.