32.3.Tweak your Form based on the Underlying Data{Accessing Data via $options}
Remove the dd() and create a new variable: $isEdit =  && $article->getId().
You might think that it's enough just to check whether $article is an object. But actually, on our new endpoint, if we wanted, we could instantiate a new Article() object and pass it as the second argument to createForm(). You do this sometimes if you want to pre-fill a "new" form with some default data. The form system would update that Article object, but Doctrine would still be smart enough to insert a new row when we save.
Anyways, that's why I'm checking not only that the Article is an object, but that it also has an id.