32.4.Tweak your Form based on the Underlying Data{Dynamically disabling a Field}
Dynamically disabling a Field
This is great, because, our goal was to disable the author field on the edit form. To do that, we can take advantage of an option that every field type has: disabled. Set it to $isEdit.
Ok, let's try that out! Refresh the edit page. Disabled! Now try the new page: not disabled. Perfect!
Oh, by the way, this disabled option does two things. First, obviously, it adds a disabled attribute so that the browser prevents the user from modifying it. But it also now ignores any submitted data for this field. So, if a nasty user removed the disabled attribute and updated the field, meh - no problem - our form will ignore that submitted data.