32.6.Tweak your Form based on the Underlying Data{Conditionally Hiding / Showing a Field}
So... we'll add it! Copy the option name, go into ArticleFormType and, down in configureOptions(), add include_published_at set to false. This is enough to make this a valid option... with a default value.
Now, up in buildForm(), the $options array will always have an include_published_at key. We can use that below to say if ($options['include_published_at']), then we want that field. Remove it from above, then say $builder paste and... clean that up a little bit.
I love it! On the edit form, because we've overridden that option to be true, when we refresh... yes! We have the field! Open up the profiler for your form and click on the top level. Nice! You can see that a passed option include_published_at was set to true.
For the new page, we should not have that field. Try it! Woh! An error from Twig:

  Neither the property publishedAt nor one of the methods publishedAt(), blah blah blah, exist in some FormView class.