32.7.Tweak your Form based on the Underlying Data{Conditionally Hiding / Showing a Field}
It's blowing up inside form_row() because we're trying to render a field that doesn't exist! Go open that template: templates/article_admin/_form.html.twig, and wrap this in an if statement: {% if articleForm.publishedAt is defined %}, then we'll render the field.
Try it again. The field is gone! And because it's completely gone from the form, when we submit, the form system will not call the setPublishedAt() method at all.
Next: let's talk about another approach to handling the situation where your form looks different than your entity class: data transfer objects.